[refactor] 프로젝트 상세내용안에 구현해둔 section 컴포넌트를 article 컴포넌트로 변경 - container 태그 section -> article로 변경 - 컴포넌트명 및 파일명 변경
