+++ src/app/projects/diary/page.tsx
@@ -9,10 +9,10 @@ import AddPostGif from '../../../../public/images/projects/diary/diary_add_diary
 import UpdatePostGif from '../../../../public/images/projects/diary/diary_update_diary.gif';
 import DeleteGif from '../../../../public/images/projects/diary/diary_delete_diary.gif';
 import Image from 'next/image';
-import ProjectServiceLinkList from '@/components/sections/project/ProjectServiceLinkSection';
-import ProjectUseSkillSection from '@/components/sections/project/ProjectUseSkillSection';
-import ProjectNumberOfPeopleSection from '@/components/sections/project/ProjectNumberOfPeopleSection';
-import ProjectWorkDetailSection from '@/components/sections/project/ProjectWorkDetailSection';
+import ProjectServiceLinkList from '@/components/articles/project/ProjectServiceLinkArticle';
+import ProjectUseSkillSection from '@/components/articles/project/ProjectUseSkillArticle';
+import ProjectNumberOfPeopleSection from '@/components/articles/project/ProjectNumberOfPeopleArticle';
+import ProjectWorkDetailSection from '@/components/articles/project/ProjectWorkDetailArticle';
 
 const links = [
   {
