+++ src/app/projects/siklog/page.tsx
@@ -8,10 +8,10 @@ import HomeImg from '../../../../public/images/projects/siklog/siklog_home.png';
 import PostDetailImg from '../../../../public/images/projects/siklog/siklog_post_detail.png';
 import SearchPostImg from '../../../../public/images/projects/siklog/siklog_search_post.png';
 import Image from 'next/image';
-import ProjectServiceLinkList from '@/components/sections/project/ProjectServiceLinkSection';
-import ProjectUseSkillSection from '@/components/sections/project/ProjectUseSkillSection';
-import ProjectNumberOfPeopleSection from '@/components/sections/project/ProjectNumberOfPeopleSection';
-import ProjectWorkDetailSection from '@/components/sections/project/ProjectWorkDetailSection';
+import ProjectServiceLinkList from '@/components/articles/project/ProjectServiceLinkArticle';
+import ProjectUseSkillSection from '@/components/articles/project/ProjectUseSkillArticle';
+import ProjectNumberOfPeopleSection from '@/components/articles/project/ProjectNumberOfPeopleArticle';
+import ProjectWorkDetailSection from '@/components/articles/project/ProjectWorkDetailArticle';
 
 const links = [
   {
