+++ src/app/projects/withwe-homepage/page.tsx
@@ -2,11 +2,11 @@
 
 import FadeIn from '@/components/gasp/FadeIn';
 import Underline from '@/components/gasp/UnderLine';
-import ProjectNumberOfPeopleSection from '@/components/sections/project/ProjectNumberOfPeopleSection';
-import ProjectRoleSection from '@/components/sections/project/ProjectRoleSection';
-import ProjectServiceLinkList from '@/components/sections/project/ProjectServiceLinkSection';
-import ProjectUseSkillSection from '@/components/sections/project/ProjectUseSkillSection';
-import ProjectWorkDetailSection from '@/components/sections/project/ProjectWorkDetailSection';
+import ProjectNumberOfPeopleSection from '@/components/articles/project/ProjectNumberOfPeopleArticle';
+import ProjectRoleSection from '@/components/articles/project/ProjectRoleArticle';
+import ProjectServiceLinkListSection from '@/components/articles/project/ProjectServiceLinkArticle';
+import ProjectUseSkillSection from '@/components/articles/project/ProjectUseSkillArticle';
+import ProjectWorkDetailSection from '@/components/articles/project/ProjectWorkDetailArticle';
 import React, { useEffect } from 'react';
 
 const links = [
@@ -53,7 +53,7 @@ export default function WithweHomePage() {
         <p className="mt-4">회사를 대표하는 홈페이지이며 Next.js를 활용하여 개발하였습니다.</p>
       </FadeIn>
       <FadeIn delay={0.9}>
-        <ProjectServiceLinkList links={links} />
+        <ProjectServiceLinkListSection links={links} />
       </FadeIn>
       <FadeIn delay={1.2}>
         <ProjectUseSkillSection skills={skills} />
