+++ src/components/articles/project/ProjectWorkDetailArticle.tsx
@@ -9,9 +9,9 @@ const subTitleStyle = 'mt-16 mobile:mt-8 text-2xl font-bold';
 const listStyle = 'mt-4 list-disc';
 const listIndentStyle = 'mt-1 list-disc !list-[circle] ml-6';
 
-export default function ProjectWorkDetailSection({ title, workDetails }: Props) {
+export default function ProjectWorkDetailArticle({ title, workDetails }: Props) {
   return (
-    <section>
+    <article>
       <h2 className={subTitleStyle}>{title}</h2>
       <ul className="pl-6">
         {workDetails.map(({ content, contentDetails }, i) => (
@@ -28,6 +28,6 @@ export default function ProjectWorkDetailSection({ title, workDetails }: Props) 
           </li>
         ))}
       </ul>
-    </section>
+    </article>
   );
 }
